Initial integration into the rest of the interpreter
