+++ src/vm/object_support/factory.hpp
@@ -1,6 +1,7 @@
 #ifndef TIRO_VM_OBJECT_SUPPORT_FACTORY_HPP
 #define TIRO_VM_OBJECT_SUPPORT_FACTORY_HPP
 
+#include "common/assert.hpp"
 #include "common/defs.hpp"
 #include "vm/context.hpp"
 #include "vm/object_support/layout.hpp"
@@ -9,23 +10,25 @@ namespace tiro::vm {
 
 namespace detail {
 
-template<typename Layout, typename... LayoutArgs>
-auto create_object_static(Heap& heap, Header* type, LayoutArgs&&... layout_args) {
-    return heap.create<Layout>(type, std::forward<LayoutArgs>(layout_args)...);
+template<typename Layout, typename... Args>
+inline Layout* create_impl(Heap& heap, Header* type, Args&&... args) {
+    using Traits = LayoutTraits<Layout>;
+    static_assert(
+        Traits::has_static_size, "the layout has dynamic size, use create_varsize instead");
+    return heap.template create<Layout>(Traits::static_size, type, std::forward<Args>(args)...);
 }
 
-template<typename Layout, typename SizeArg, typename... LayoutArgs>
-auto create_object_varsize(
-    Heap& heap, Header* type, SizeArg&& size_arg, LayoutArgs&&... layout_args) {
+template<typename Layout, typename SizeArg, typename... Args>
+inline Layout* create_varsize_impl(Heap& heap, Header* type, SizeArg&& size_arg, Args&&... args) {
     using Traits = LayoutTraits<Layout>;
+    static_assert(!Traits::has_static_size, "the layout has static size, use create() instead");
 
-    const size_t allocation_size = Traits::dynamic_alloc_size(size_arg);
+    const size_t total_byte_size = Traits::dynamic_alloc_size(size_arg);
 
-    auto instance = heap.create_varsize<Layout>(
-        allocation_size, type, std::forward<LayoutArgs>(layout_args)...);
-    TIRO_DEBUG_ASSERT(Traits::dynamic_size(instance) == allocation_size,
-        "Variable size object must report exactly the requested size.");
-    return instance;
+    Layout* data = heap.template create<Layout>(total_byte_size, type, std::forward<Args>(args)...);
+    TIRO_DEBUG_ASSERT(Traits::dynamic_size(data) == total_byte_size,
+        "byte size mismatch between requested and calculated dynamic object size");
+    return data;
 }
 
 } // namespace detail
@@ -36,10 +39,10 @@ auto create_object(Context& ctx, LayoutArgs&&... layout_args) {
 
     auto type = ctx.types().raw_internal_type<BuiltinType>(); // rooted by the TypeSystem instance
     if constexpr (LayoutTraits<Layout>::has_static_size) {
-        return detail::create_object_static<Layout>(
+        return detail::create_impl<Layout>(
             ctx.heap(), type, std::forward<LayoutArgs>(layout_args)...);
     } else {
-        return detail::create_object_varsize<Layout>(
+        return detail::create_varsize_impl<Layout>(
             ctx.heap(), type, std::forward<LayoutArgs>(layout_args)...);
     }
 }
