+++ test/unit_tests/vm/heap/CMakeLists.txt
@@ -1,8 +1,5 @@
 target_sources(unit_tests
     PRIVATE
-        collector_test.cpp
         heap_test.cpp
         memory_test.cpp
-
-        new_heap_test.cpp
 )
