+++ test/unit_tests/vm/heap/OLD_heap_test.cpp
@@ -1,0 +1,75 @@
+#include <catch2/catch.hpp>
+
+#include "vm/heap/heap.hpp"
+#include "vm/objects/value.hpp"
+
+namespace tiro::vm::test {
+
+TEST_CASE("Object list should be able to traverse the heap", "[heap]") {
+    Header h1{Header::InvalidTag()}, h2{Header::InvalidTag()}, h3{Header::InvalidTag()};
+
+    ObjectList list;
+    list.insert(TIRO_NN(&h1));
+    list.insert(TIRO_NN(&h2));
+    list.insert(TIRO_NN(&h3));
+    REQUIRE_FALSE(list.empty());
+
+    SECTION("simple iteration") {
+        ObjectList::Cursor cursor = list.cursor();
+        int index = 0;
+        while (cursor) {
+            CAPTURE(index);
+
+            Header* obj = cursor.get();
+            REQUIRE(obj);
+
+            switch (index) {
+            case 0:
+                REQUIRE(obj == &h3);
+                break;
+            case 1:
+                REQUIRE(obj == &h2);
+                break;
+            case 2:
+                REQUIRE(obj == &h1);
+                break;
+            }
+
+            cursor.next();
+            ++index;
+        }
+
+        REQUIRE(index == 3);
+    }
+
+    SECTION("remove all") {
+        ObjectList::Cursor cursor = list.cursor();
+        int index = 0;
+        while (cursor) {
+            CAPTURE(index);
+
+            Header* obj = cursor.get();
+            REQUIRE(obj);
+
+            switch (index) {
+            case 0:
+                REQUIRE(obj == &h3);
+                break;
+            case 1:
+                REQUIRE(obj == &h2);
+                break;
+            case 2:
+                REQUIRE(obj == &h1);
+                break;
+            }
+
+            cursor.remove();
+            ++index;
+        }
+
+        REQUIRE(index == 3);
+        REQUIRE(list.empty());
+    }
+}
+
+} // namespace tiro::vm::test
